+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/components/index.ts
@@ -1,9 +1,9 @@
 export { ControlBar } from './control-bar/control-bar'
 export { ErrorBoundary } from './error/index'
-export { LoopNodeComponent } from './loop-node/loop-node'
 export { Panel } from './panel/panel'
-export { ParallelNodeComponent } from './parallel-node/parallel-node'
 export { SkeletonLoading } from './skeleton-loading/skeleton-loading'
+export { LoopNodeComponent } from './subflows/loop/loop-node'
+export { ParallelNodeComponent } from './subflows/parallel/parallel-node'
 export { WandPromptBar } from './wand-prompt-bar/wand-prompt-bar'
 export { WorkflowBlock } from './workflow-block/workflow-block'
 export { WorkflowEdge } from './workflow-edge/workflow-edge'
